2020.07.14

+++ drawtune.c
@@ -1475,7 +1475,7 @@ static void sizevoice(struct voice* v, struct tune* t)
     case DYNAMIC:
       break;
     case LINENUM:
-      lineno = (int)(ft->item);
+      lineno = (long)(ft->item);
       break;
     case MUSICLINE: 
       break;
@@ -2958,7 +2958,7 @@ static int printvoiceline(struct voice* v)
           (v->place->type == LEFT_TEXT) || (v->place->type == CENTRE_TEXT) ||
           (v->place->type == VSKIP))) {
     if (v->place->type == LINENUM) {
-      lineno = (int)(v->place->item);
+      lineno = (long)(v->place->item);
     };
     if (v->place->type == NEWPAGE) {
       newpage();
@@ -2970,7 +2970,7 @@ static int printvoiceline(struct voice* v)
       printtext(centre, v->place->item, &textfont);
     };
     if (v->place->type == VSKIP) {
-      vskip((double)((int)v->place->item));
+      vskip((double)((long)v->place->item));
     };
     v->place = v->place->next;
   };
@@ -3004,21 +3004,21 @@ static int printvoiceline(struct voice* v)
     switch (ft->type) {
     case SINGLE_BAR: 
       fprintf(f, "%.1f bar\n", ft->x);
-      printbarnumber(ft->x, (int)ft->item);
+      printbarnumber(ft->x, (long)ft->item);
       break;
     case DOUBLE_BAR: 
       fprintf(f, "%.1f dbar\n", ft->x);
-      printbarnumber(ft->x, (int)ft->item);
+      printbarnumber(ft->x, (long)ft->item);
       inend = endrep(inend, endstr, xend, ft->x, spacing->yend);
       break;
     case BAR_REP: 
       fprintf(f, "%.1f fbar1 %.1f rdots\n", ft->x, ft->x+10);
-      printbarnumber(ft->x, (int)ft->item);
+      printbarnumber(ft->x, (long)ft->item);
       inend = endrep(inend, endstr, xend, ft->x, spacing->yend);
       break;
     case REP_BAR: 
       fprintf(f, "%.1f rdots %.1f fbar2\n", ft->x, ft->x+10);
-      printbarnumber(ft->x, (int)ft->item);
+      printbarnumber(ft->x, (long)ft->item);
       inend = endrep(inend, endstr, xend, ft->x, spacing->yend);
       break;
     case REP1: 
@@ -3041,7 +3041,7 @@ static int printvoiceline(struct voice* v)
       break;
     case BAR1: 
       fprintf(f, "%.1f bar\n", ft->x);
-      printbarnumber(ft->x, (int)ft->item);
+      printbarnumber(ft->x, (long)ft->item);
       inend = endrep(inend, endstr, xend, ft->x - ft->xleft, spacing->yend);
       inend = 1;
       strcpy(endstr, "1");
@@ -3049,7 +3049,7 @@ static int printvoiceline(struct voice* v)
       break;
     case REP_BAR2: 
       fprintf(f, "%.1f rdots %.1f fbar2\n", ft->x, ft->x+10);
-      printbarnumber(ft->x, (int)ft->item);
+      printbarnumber(ft->x, (long)ft->item);
       inend = endrep(inend, endstr, xend, ft->x - ft->xleft, spacing->yend);
       inend = 2;
       strcpy(endstr, "2");
@@ -3242,7 +3242,7 @@ static int printvoiceline(struct voice* v)
       if(psaction->color == 'b') redcolor = 0;
       break;
     case LINENUM: 
-      lineno = (int)(ft->item);
+      lineno = (long)(ft->item);
       break;
     case MUSICLINE: 
       v->line = midline;
@@ -3324,7 +3324,7 @@ static int finalsizeline(struct voice* v)
     ft = ft->next;
   };
   if ((ft != NULL) && (ft->type == PRINTLINE)) {
-    avertspacing = ft->item;
+    avertspacing =  ft->item;
     avertspacing->height = (float) height;
     avertspacing->descender = (float) descender;
     avertspacing->yend = (float) yend;
@@ -3365,7 +3365,7 @@ static int getlineheight(struct voice* v, double* height)
           (v->place->type == LEFT_TEXT) || (v->place->type == CENTRE_TEXT) ||
           (v->place->type == VSKIP))) {
     if (v->place->type == LINENUM) {
-      lineno = (int)(v->place->item);
+      lineno = (long)(v->place->item);
     };
     if (v->place->type == LEFT_TEXT) {
       *height = *height + textfont.pointsize + textfont.space;
@@ -3374,7 +3374,7 @@ static int getlineheight(struct voice* v, double* height)
       *height = *height + textfont.pointsize + textfont.space;
     };
     if (v->place->type == VSKIP) {
-      *height = *height + (double)((int)v->place->item);
+      *height = *height + (double)((long)v->place->item);
     };
     v->place = v->place->next;
   };
